Create FLStudioSortUnknownVendorUsingFuzzyLogic
This is not an optimized script still experimental - the idea of it is simply to allow for the unknown folder at documents/VENDORLIST/UNKNOWN to be sorted with minimal fuzzy logic in trying to guess what the plugin vendor is if it isn't indicated from the .nfo file. It sort of works but very crudely.