Create findandmovflpwavfilestofolder-v1.ps1
Tbe purpose of this script is to let you select the flp file to search for .wav file search the location and if it doesn't find it to search alternate search locations, you can also designate a drive. It searches any file path that is found first. This is the first version. Due to flp files being partially encoded  not all file paths or names may be found in this version. Lets the user select an .flp file and an output folder.

Extracts potential .wav file paths from the FLP.

Expands variables like %FLStudioFactoryData%.

Searches prioritized folders and, if needed, the full C:\ drive.

Displays color-coded results in a custom ListBox.

Shows real-time status updates.